requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
python-dotenv==1.0.0
ta==0.11.0
tqdm==4.65.0
joblib==1.3.2
torch>=2.0.1
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17
torch-cluster>=1.6.1
torch-spline-conv>=1.2.2
networkx>=3.0
statsmodels>=0.14.0
scipy>=1.10.0
numba>=0.57.0